Move CodexAuth and AuthManager to the core crate (#3074)
Fix a long standing layering issue.

+++ codex-rs/tui/src/bottom_pane/chat_composer.rs
@@ -2042,7 +2042,7 @@ mod tests {
                 composer.handle_paste(paste.clone());
                 composer
                     .textarea
-                    .set_cursor((placeholder.len() - pos_from_end) as usize);
+                    .set_cursor(placeholder.len() - pos_from_end);
                 composer.handle_key_event(KeyEvent::new(KeyCode::Backspace, KeyModifiers::NONE));
                 let result = (
                     composer.textarea.text().contains(&placeholder),
